Fix HTTP status

+++ backend/src/main/java/de/neuefische/cgnjava222/zoomanagement/zoo/animal/AnimalService.java
@@ -27,7 +27,12 @@ public class AnimalService {
 
     }
 
-    public void deleteAnimal(String id) {
-        animalRepo.deleteById(id);
+    public boolean deleteAnimal(String id) {
+        if (animalRepo.existsById(id)) {
+            animalRepo.deleteById(id);
+            return true;
+        }
+        return false;
     }
+
 }
